backend/requirements.txt
# Core FastAPI dependencies
fastapi
uvicorn
pydantic
python-dotenv

# Database
SQLAlchemy

